+++ CHANGES.rst
@@ -1856,6 +1856,9 @@ astropy.vo
 astropy.wcs
 ^^^^^^^^^^^
 
+- Fix bug that caused ``WCS.has_celestial`` and other functionality depending on
+  it to fail in the presence of correlated celestial and other axes. [#8420]
+
 
 Other Changes and Additions
 ---------------------------